add some units test

+++ src/Unit/Bund.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Bund extends Unit
 {
-    protected int $conversionFactor = 1;
+    protected float $conversionFactor = 1;
     protected string $referenceUnitClass = Bund::class;
     protected string $shortCode = 'Bd';
