+++ src/Unit/CupUK.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class CupUK extends Unit
 {
-    protected int $conversionFactor = 250;
+    protected float $conversionFactor = 250;
     protected string $referenceUnitClass = Mililiter::class;
     protected string $shortCode = 'c';
