+++ src/Unit/CupUS.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class CupUS extends Unit
 {
-    protected int $conversionFactor = 238;
+    protected float $conversionFactor = 238;
     protected string $referenceUnitClass = Mililiter::class;
     protected string $shortCode = 'c';
 }
