+++ src/Unit/Deziliter.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Deziliter extends Unit
 {
-    protected int $conversionFactor = 1/100;
+    protected float $conversionFactor = 1/100;
     protected string $referenceUnitClass = Mililiter::class;
     protected string $shortCode = 'dl';
 }
