+++ src/Unit/Grain.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Grain extends Unit
 {
-    protected int $conversionFactor = 65/1000;
+    protected float $conversionFactor = 65/1000;
     protected string $referenceUnitClass = Gramm::class;
     protected string $shortCode = 'gr';
