+++ src/Unit/Gramm.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Gramm extends Unit
 {
-    protected int $conversionFactor = 1;
+    protected float $conversionFactor = 1;
     protected string $referenceUnitClass = Gramm::class;
     protected string $shortCode = 'g';
 }
