+++ src/Unit/Milligramm.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Milligramm extends Unit
 {
-    protected int $conversionFactor = 10;
+    protected float $conversionFactor = 10;
     protected string $referenceUnitClass = Gramm::class;
     protected string $shortCode = 'mg';
 }
