+++ src/Unit/Ounce.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Ounce extends Unit
 {
-    protected int $conversionFactor = 2835/100;
+    protected float $conversionFactor = 2835/100;
     protected string $referenceUnitClass = Gramm::class;
     protected string $shortCode = 'oz.';
