+++ src/Unit/Pint.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Pint extends Unit
 {
-    protected int $conversionFactor = 470;
+    protected float $conversionFactor = 470;
     protected string $referenceUnitClass = Mililiter::class;
     protected string $shortCode = 'pt.';
 }
