+++ src/Unit/Pound.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Pound extends Unit
 {
-    protected int $conversionFactor = 454;
+    protected float $conversionFactor = 454;
     protected string $referenceUnitClass = Gramm::class;
     protected string $shortCode = 'lb.';
