+++ src/Unit/Quart.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Quart extends Unit
 {
-    protected int $conversionFactor = 950;
+    protected float $conversionFactor = 950;
     protected string $referenceUnitClass = Mililiter::class;
     protected string $shortCode = 'qt.';
 }
