+++ src/Unit/SaltSpoont.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class SaltSpoont extends Unit
 {
-    protected int $conversionFactor = 1;
+    protected float $conversionFactor = 1;
     protected string $referenceUnitClass = Mililiter::class;
     protected string $shortCode = 'ssp.';
