+++ src/Unit/Spritzer.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Spritzer extends Unit
 {
-    protected int $conversionFactor = 4;
+    protected float $conversionFactor = 4;
     protected string $referenceUnitClass = Tropfen::class;
     protected string $shortCode = 'Tr';
 }
