+++ src/Unit/Tasse.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Tasse extends Unit
 {
-    protected int $conversionFactor = 200;
+    protected float $conversionFactor = 200;
     protected string $referenceUnitClass = Mililiter::class;
     protected string $shortCode = 'Tas';
 }
