+++ src/Unit/TeaCup.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class TeaCup extends Unit
 {
-    protected int $conversionFactor = 190;
+    protected float $conversionFactor = 190;
     protected string $referenceUnitClass = Mililiter::class;
     protected string $shortCode = 'tc';
