+++ src/Unit/Teeloeffel.php
@@ -6,7 +6,7 @@ namespace OrdbModels\Unit;
 
 class Teeloeffel extends Unit
 {
-    protected int $conversionFactor = 5;
+    protected float $conversionFactor = 5;
     protected string $referenceUnitClass = Mililiter::class;
     protected string $shortCode = 'TL';
